Remove RRC questions and references from puerto rico intake
