+++ app/controllers/ctc/questions/main_home_controller.rb
@@ -10,7 +10,7 @@ module Ctc
       def next_path
         if current_intake.home_location_us_territory? || current_intake.home_location_foreign_address?
           questions_use_gyr_path
-        elsif current_intake.home_location_puerto_rico?
+        elsif current_intake.home_location_puerto_rico? && !current_intake.puerto_rico_filing?
           offboarding_cant_use_getctc_pr_path
         else
           super
