+++ app/controllers/ctc/questions/stimulus_received_controller.rb
@@ -6,6 +6,7 @@ module Ctc
       layout "intake"
 
       def self.show?(intake)
+        return false if intake.puerto_rico_filing?
         return false if intake.eip3_amount_received.nil?
 
         Efile::BenefitsEligibility.new(tax_return: intake.default_tax_return, dependents: intake.dependents).outstanding_recovery_rebate_credit&.zero?
