+++ app/views/ctc/questions/confirm_payment/edit.html.erb
@@ -10,11 +10,12 @@
       <p class="spacing-below-15"><%= number_to_currency(@ctc_amount) %></p>
       <%= link_to t('views.ctc.questions.confirm_payment.ctc_0_due_link'), Ctc::Questions::Dependents::InfoController.to_path_helper(id: current_intake.new_dependent_token), class: "link" if @ctc_amount.zero? %>
     </div>
-
-    <div class="spacing-above-35">
-      <p class="spacing-below-15"><strong><%= t('views.ctc.questions.confirm_payment.third_stimulus') %></strong></p>
-      <p class="spacing-below-15"><%= number_to_currency(@third_stimulus_amount) %></p>
-    </div>
+    <% unless current_intake.puerto_rico_filing? %>
+      <div class="spacing-above-35">
+        <p class="spacing-below-15"><strong><%= t('views.ctc.questions.confirm_payment.third_stimulus') %></strong></p>
+        <p class="spacing-below-15"><%= number_to_currency(@third_stimulus_amount) %></p>
+      </div>
+    <% end %>
   </div>
 
   <%= link_to t("general.continue"), next_path, class: "button button--primary button--wide text--centered spacing-above-60" unless @not_collecting%>
