+++ config/locales/es.yml
@@ -1944,6 +1944,7 @@ es:
                 title: "¿Debo declarar como jefe de familia?"
               p1_html: Si obtuvo ingresos en %{current_tax_year}, uno o más de sus dependientes podrían calificarlo para un Crédito Tributario por Ingreso del Trabajo más grande. El EITC está disponible para algunas personas que obtienen ingresos, y es mayor para aquellas que tienen hijos. Desafortunadamente, usted actualmente no puede reclamar el EITC en la declaración simplificada que presenta con GetCTC. Para presentar una declaración de impuestos completa y reclamar el EITC, visite <a href="https://getyourrefund.org/gctc_ref">GetYourRefund</a>.
             title: "¿Qué pasa con otros beneficios para los que podrían ser elegibles mis dependientes?"
+          qualifying_for_both: Dependientes para el CTC
           qualifying_for_ctc: Dependientes para el CTC y el pago de estímulo
           qualifying_for_other_credits: Dependientes con sólo derecho a recibir el pago de estímulo
           ssn: Número de Seguro Social
