+++ spec/lib/efile/dependent_eligibility/eip_three_spec.rb
@@ -1,6 +1,13 @@
 require "rails_helper"
 
 describe Efile::DependentEligibility::EipThree do
+  context "when puerto_rico_filing? is true on the intake" do
+    it "returns false, because PR dependents can not get eip3 money" do
+      eligibility = described_class.new((create :qualifying_child, intake: create(:ctc_intake, home_location: "puerto_rico", client: create(:client, :with_return))), TaxReturn.current_tax_year)
+      expect(eligibility.qualifies?).to eq false
+    end
+  end
+
   context "when passing in a qualifying child object" do
     let(:child_eligibility) { double }
     before do
@@ -11,7 +18,6 @@ describe Efile::DependentEligibility::EipThree do
       described_class.new((create :dependent), TaxReturn.current_tax_year, child_eligibility: child_eligibility)
       expect(child_eligibility).to have_received(:qualifies?)
     end
-
   end
 
   context "when passing in a qualifying relative object" do
@@ -24,6 +30,5 @@ describe Efile::DependentEligibility::EipThree do
       described_class.new((create :dependent, intake: create(:ctc_intake, client: create(:client, :with_return))), TaxReturn.current_tax_year, relative_eligibility: relative_eligibility)
       expect(relative_eligibility).to have_received(:qualifies?)
     end
-
   end
 end
